Refactor dashboard server modules
